# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Plant Care Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GrowthLog {
  id: number;
  plantId: number;
  date: string;
  notes: string;
  /** @nullable */
  photo?: string | null;
}

export interface Plant {
  id: number;
  name: string;
  species: string;
  /** @nullable */
  photo?: string | null;
  dateAcquired: string;
  wateringFrequencyDays: number;
  /** @nullable */
  lastWatered?: string | null;
  logs: GrowthLog[];
}

export interface PlantInput {
  /** @minLength 1 */
  name: string;
  species: string;
  photo?: string;
  dateAcquired: string;
  /** @minimum 1 */
  wateringFrequencyDays: number;
  lastWatered?: string;
}

export interface PlantUpdate {
  /** @minLength 1 */
  name?: string;
  species?: string;
  /** @nullable */
  photo?: string | null;
  dateAcquired?: string;
  /** @minimum 1 */
  wateringFrequencyDays?: number;
}

export interface GrowthLogInput {
  /** @minLength 1 */
  notes: string;
  photo?: string;
  date?: string;
}

export type PlantHealthStatus = typeof PlantHealthStatus[keyof typeof PlantHealthStatus];


export const PlantHealthStatus = {
  good: 'good',
  'due-soon': 'due-soon',
  overdue: 'overdue',
} as const;

export interface PlantHealth {
  plantId: number;
  name: string;
  status: PlantHealthStatus;
  /** @nullable */
  daysUntilWatering: number | null;
  wateringFrequencyDays: number;
}

export interface GardenStats {
  totalPlants: number;
  overdueCount: number;
  dueSoonCount: number;
  healthyCount: number;
  totalLogs: number;
  wateredThisWeek: number;
  careStreak: number;
  plantHealth: PlantHealth[];
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  watered: 'watered',
  log: 'log',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  plantId: number;
  plantName: string;
  date: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  photo?: string | null;
}