# lib/api-zod/src/generated/types/gardenStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Plant Care Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { PlantHealth } from './plantHealth';

export interface GardenStats {
  totalPlants: number;
  overdueCount: number;
  dueSoonCount: number;
  healthyCount: number;
  totalLogs: number;
  wateredThisWeek: number;
  careStreak: number;
  plantHealth: PlantHealth[];
}