# lib/api-zod/src/generated/types/growthLog.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Plant Care Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface GrowthLog {
  id: number;
  plantId: number;
  date: Date;
  notes: string;
  /** @nullable */
  photo?: string | null;
}