# lib/api-zod/src/generated/types/growthLogInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Plant Care Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface GrowthLogInput {
  /** @minLength 1 */
  notes: string;
  photo?: string;
  date?: Date;
}