# lib/api-zod/src/generated/types/plant.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Plant Care Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { GrowthLog } from './growthLog';

export interface Plant {
  id: number;
  name: string;
  species: string;
  /** @nullable */
  photo?: string | null;
  dateAcquired: Date;
  wateringFrequencyDays: number;
  /** @nullable */
  lastWatered?: Date | null;
  logs: GrowthLog[];
}