# lib/api-zod/src/generated/types/plantHealth.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Plant Care Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { PlantHealthStatus } from './plantHealthStatus';

export interface PlantHealth {
  plantId: number;
  name: string;
  status: PlantHealthStatus;
  /** @nullable */
  daysUntilWatering: number | null;
  wateringFrequencyDays: number;
}