# lib/api-zod/src/generated/types/plantInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Plant Care Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface PlantInput {
  /** @minLength 1 */
  name: string;
  species: string;
  photo?: string;
  dateAcquired: Date;
  /** @minimum 1 */
  wateringFrequencyDays: number;
  lastWatered?: Date;
}