# lib/api-zod/src/generated/types/plantUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Plant Care Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface PlantUpdate {
  /** @minLength 1 */
  name?: string;
  species?: string;
  /** @nullable */
  photo?: string | null;
  dateAcquired?: Date;
  /** @minimum 1 */
  wateringFrequencyDays?: number;
}